+++ src/components/common/MainBanner/MainBanner.style.ts
